requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-cohere>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-core>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0
streamlit-chat>=0.1.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
cohere>=4.0.0
typing-extensions>=4.5.0
